update: packer.nvim -> lazy.nvim
